requirements.txt
async-timeout==4.0.3
blinker==1.6.2
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.12.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///C:/ci/conda-package-handling_1649106011304/work
cryptography @ file:///C:/ci/cryptography_1639414720302/work
filelock==3.12.4
Flask==2.3.3
gunicorn==21.2.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
menuinst @ file:///C:/ci/menuinst_1631733438520/work
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper==20230918
packaging==23.1
Pillow==9.3.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dotenv==1.0.0
pywin32==302
redis==5.0.1
regex==2023.8.8
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
spotipy==2.23.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
typing_extensions==4.8.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
Werkzeug==2.3.7
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
zipp==3.17.0
